service/pyproject.toml
[project]
name = "roundtable-api"
version = "0.1.0"
description = "FastAPI backend for the Crucible platform."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "upstash-redis>=1.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.0",
    "argon2-cffi>=23.0.0",
    "playwright>=1.48.0",
    "python-jose[cryptography]>=3.3.0",
    "structlog>=24.1.0",
    "orjson>=3.9.0",
    "pymupdf>=1.24.8",
    "openai>=1.0.0",
    "sse-starlette>=1.8.0",
    "json-repair>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "coverage[toml]>=7.4.0",
    "ruff>=0.5.5",
    "mypy>=1.10.0",
    "types-redis>=4.6.0.3",
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
